latest

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,13 +10,13 @@
 int main(void)
 {
 	int count = 2;
-	long long i = 1;
-	long long j = 2;
+	unsigned long i = 1;
+	unsigned long j = 2;
 
 	printf("1, 2, ");
 	while (count != 98)
 	{
-		long long new = i + j;
+		unsigned long new = i + j;
 
 		i = j;
 		j = new;
